refactor: 13-ref2vec folder structure
